refactor(*): LottoController 추가
